week_8/requirements.txt
streamlit
groq
langchain-mcp-adapters
langchain
httpx
python-dotenv
